config: chromiumos: don't rebuild sdk toolchain
The CrOS SDK tarball already contains a precompiled toolchain
ready for use and the sdk just rebuilds it from source by
default because the system is not identical to the one which
created the tarball.

Considering CrOS does not update the toolchains in stable
versions, and that KCI is only using stable releases, skip
the chroot toolchain rebuild to save some precious minutes
in each board build.

Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>